refactor: split compose.py into modules
