+++ scripts/compose.py
@@ -1,26 +1,22 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 """
 CLI for starting a testing environment using docker-compose.
 """
 from __future__ import print_function
 
-from abc import abstractmethod
-
 import argparse
-import codecs
 import collections
-from collections import OrderedDict
 import datetime
-import functools
 import glob
-import inspect
-import json
 import logging
-import multiprocessing
-import os
-import re
-import sys
-import subprocess
+import inspect
+
+from scripts.aux_services import *
+from scripts.elastic_stack import *
+from scripts.beats import *
+from scripts.opbeans import *
+from scripts.service import *
+from scripts.helpers import *
 
 try:
     from urllib.request import urlopen, urlretrieve, Request
@@ -34,17 +30,8 @@ except ImportError:
 PACKAGE_NAME = 'localmanager'
 __version__ = "4.0.0"
 
-DEFAULT_STACK_VERSION = "8.0"
+
 DEFAULT_APM_SERVER_URL = "http://apm-server:8200"
-DEFAULT_APM_JS_SERVER_URL = "http://localhost:8200"
-
-
-#
-# helpers
-#
-def _camel_hyphen(string):
-    return re.sub(r'([a-z])([A-Z])', r'\1-\2', string)
-
 
 def discover_services(mod=None):
     """discover list of services"""
@@ -57,2354 +44,6 @@ def discover_services(mod=None):
                 and cls not in (Service, OpbeansService):
             ret.append(cls)
     return ret
-
-
-def _load_image(cache_dir, url):
-    filename = os.path.basename(url)
-    filepath = os.path.join(cache_dir, filename)
-    etag_cache_file = filepath + '.etag'
-    if os.path.exists(etag_cache_file):
-        with open(etag_cache_file, mode='r') as f:
-            etag = f.read().strip()
-    else:
-        etag = None
-    request = Request(url)
-    request.get_method = lambda: 'HEAD'
-    try:
-        response = urlopen(request)
-    except Exception as e:
-        print('Error while fetching %s: %s' % (url, str(e)))
-        return False
-    new_etag = response.info().get('ETag')
-    if etag == new_etag:
-        print("Skipping download of %s, local file is current" % filename)
-        return True
-    print("downloading", url)
-    try:
-        os.makedirs(cache_dir)
-    except Exception:  # noqa: E722
-        pass  # ignore
-    try:
-        urlretrieve(url, filepath)
-    except Exception as e:
-        print('Error while fetching %s: %s' % (url, str(e)))
-        return False
-    subprocess.check_call(["docker", "load", "-i", filepath])
-    with open(etag_cache_file, mode='w') as f:
-        f.write(new_etag)
-    return True
-
-
-def load_images(urls, cache_dir):
-    load_image_fn = functools.partial(_load_image, cache_dir)
-    pool = multiprocessing.Pool(4)
-    # b/c python2
-    try:
-        results = pool.map_async(load_image_fn, urls).get(timeout=10000000)
-    except KeyboardInterrupt:
-        pool.terminate()
-        raise
-    if not all(results):
-        print("Errors while downloading. Exiting.")
-        sys.exit(1)
-
-
-DEFAULT_HEALTHCHECK_INTERVAL = "10s"
-DEFAULT_HEALTHCHECK_RETRIES = 12
-
-
-def curl_healthcheck(port, host="localhost", path="/healthcheck",
-                     interval=DEFAULT_HEALTHCHECK_INTERVAL, retries=DEFAULT_HEALTHCHECK_RETRIES):
-    return {
-        "interval": interval,
-        "retries": retries,
-        "test": ["CMD", "curl", "--write-out", "'HTTP %{http_code}'", "--fail", "--silent",
-                 "--output", "/dev/null",
-                 "http://{}:{}{}".format(host, port, path)]
-    }
-
-
-build_manifests = {}  # version -> manifest cache
-
-
-def latest_build_manifest(version):
-    minor_version = ".".join(version.split(".", 2)[:2])
-    rsp = urlopen("https://staging.elastic.co/latest/{}.json".format(minor_version))
-    if rsp.code != 200:
-        raise Exception("failed to query build candidates at {}: {}".format(rsp.geturl(), rsp.info()))
-    encoding = "utf-8"  # python2 rsp.headers.get_content_charset("utf-8")
-    info = json.load(codecs.getreader(encoding)(rsp))
-    if "summary_url" in info:
-        print("found latest build candidate for {} - {} at {}".format(minor_version, info["summary_url"], rsp.geturl()))
-    return info["manifest_url"]
-
-
-def resolve_bc(version, build_id):
-    """construct or discover build candidate manifest url"""
-    if build_id is None:
-        return
-
-    if version is None:
-        return
-
-    # check cache
-    if version in build_manifests:
-        return build_manifests[version]
-
-    if build_id == "latest":
-        manifest_url = latest_build_manifest(version)
-    else:
-        manifest_url = "https://staging.elastic.co/{patch_version}-{sha}/manifest-{patch_version}.json".format(
-            patch_version=version,
-            sha=build_id,
-        )
-    rsp = urlopen(manifest_url)
-    if rsp.code != 200:
-        raise Exception("failed to fetch build manifest at {}: {}".format(rsp.geturl(), rsp.info()))
-    encoding = "utf-8"  # python2 rsp.headers.get_content_charset("utf-8")
-    manifest = json.load(codecs.getreader(encoding)(rsp))
-    build_manifests[version] = manifest  # fill cache
-    return manifest
-
-
-def parse_version(version):
-    res = []
-    for x in version.split('.'):
-        try:
-            y = int(x)
-        except ValueError:
-            y = int(x.split("-", 1)[0])
-        res.append(y)
-    return res
-
-
-def add_agent_environment(mappings):
-    def fn(func):
-        def add_content(self):
-            content = func(self)
-            for option, envvar in sorted(mappings):
-                val = self.options.get(option)
-                if val is not None:
-                    if isinstance(content["environment"], dict):
-                        content["environment"][envvar] = val
-                    else:
-                        content["environment"].append(envvar + "=" + val)
-            return content
-        return add_content
-    return fn
-
-
-class Service(object):
-    """encapsulate docker-compose service definition"""
-
-    DEFAULT_ELASTICSEARCH_HOSTS = "elasticsearch:9200"
-    DEFAULT_KIBANA_HOST = "kibana:5601"
-
-    # is this a side car service for opbeans. If yes, it will automatically
-    # start if any opbeans service starts
-    opbeans_side_car = False
-
-    def __init__(self, **options):
-        self.options = options
-
-        if not hasattr(self, "docker_registry"):
-            self.docker_registry = "docker.elastic.co"
-        if not hasattr(self, "docker_name"):
-            self.docker_name = self.name()
-        if not hasattr(self, "docker_path"):
-            self.docker_path = self.name()
-
-        if hasattr(self, "SERVICE_PORT"):
-            self.port = options.get(self.option_name() + "_port", self.SERVICE_PORT)
-
-        self._oss = options.get(self.option_name() + "_oss") or options.get("oss")
-        self._release = options.get(self.option_name() + "_release") or options.get("release")
-        self._snapshot = options.get(self.option_name() + "_snapshot") or options.get("snapshot")
-
-        # version is service specific or stack or default
-        self._version = options.get(self.option_name() + "_version") or options.get("version", DEFAULT_STACK_VERSION)
-
-        # bc depends on version for resolution
-        if not self.option_name().startswith("opbeans"):
-            self._bc = resolve_bc(self._version, options.get(self.option_name() + "_bc") or options.get("bc"))
-        else:
-            self._bc = ""
-
-        self.depends_on = {}
-
-    @property
-    def bc(self):
-        return self._bc
-
-    def default_container_name(self):
-        return "_".join(("localtesting", self.version, self.name()))
-
-    def default_image(self, version_override=None):
-        """default container image path constructor"""
-        image = "/".join((self.docker_registry, self.docker_path, self.docker_name))
-        if self.oss:
-            image += "-oss"
-        image += ":" + (version_override or self.version)
-        # no command line option for setting snapshot, snapshot == no bc and not release
-        if self.snapshot or not (any((self.bc, self.release))):
-            image += "-SNAPSHOT"
-        return image
-
-    def default_labels(self):
-        return ["co.elastic.apm.stack-version=" + self.version]
-
-    @staticmethod
-    def default_logging():
-        return {
-            "driver": "json-file",
-            "options": {
-                "max-file": "5",
-                "max-size": "2m"
-            }
-        }
-
-    @staticmethod
-    def enabled():
-        return False
-
-    def at_least_version(self, target):
-        return parse_version(self.version) >= parse_version(target)
-
-    @classmethod
-    def name(cls):
-        return _camel_hyphen(cls.__name__).lower()
-
-    @classmethod
-    def option_name(cls):
-        return cls.name().replace("-", "_")
-
-    @property
-    def oss(self):
-        return self._oss
-
-    @staticmethod
-    def publish_port(external, internal, expose=False):
-        addr = "" if expose else "127.0.0.1:"
-        return addr + ":".join((str(external), str(internal)))
-
-    @property
-    def release(self):
-        return self._release
-
-    @property
-    def snapshot(self):
-        return self._snapshot
-
-    def render(self):
-        content = self._content()
-        content.update(dict(
-            container_name=content.get("container_name", self.default_container_name()),
-            image=content.get("image", self.default_image()),
-            labels=content.get("labels", self.default_labels()),
-            logging=content.get("logging", self.default_logging())
-        ))
-        for prune in "image", "labels", "logging":
-            if content[prune] is None:
-                del (content[prune])
-
-        return {self.name(): content}
-
-    @property
-    def version(self):
-        return self._version
-
-    @classmethod
-    def add_arguments(cls, parser):
-        """add service-specific command line arguments"""
-        # allow port overrides
-        if hasattr(cls, 'SERVICE_PORT'):
-            parser.add_argument(
-                '--' + cls.name() + '-port',
-                type=int,
-                default=cls.SERVICE_PORT,
-                dest=cls.option_name() + '_port',
-                help="service port"
-            )
-
-    def image_download_url(self):
-        pass
-
-    @abstractmethod
-    def _content(self):
-        pass
-
-
-class StackService(object):
-    """Mix in for Elastic services that have public docker images built but not available in a registry [yet]"""
-
-    def build_candidate_manifest(self):
-        version = self.version
-        image = self.docker_name
-        if self.oss:
-            image += "-oss"
-        key = "{image}-{version}-docker-image.tar.gz".format(
-            image=image,
-            version=version,
-        )
-        try:
-            return self.bc["projects"][self.docker_name]["packages"][key]
-        except KeyError:
-            # help debug manifest issues
-            print(json.dumps(self.bc))
-            raise
-
-    def image_download_url(self):
-        # Elastic releases are public
-        if self.release or not self.bc:
-            return
-
-        info = self.build_candidate_manifest()
-        assert info["type"] == "docker"
-        return info["url"]
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(StackService, cls).add_arguments(parser)
-        for image_detail_key in ("bc", "version"):
-            parser.add_argument(
-                "--" + cls.name() + "-" + image_detail_key,
-                type=str,
-                dest=cls.option_name() + "_" + image_detail_key,
-                help="stack {} override".format(image_detail_key),
-            )
-        for image_detail_key in ("oss", "release", "snapshot"):
-            parser.add_argument(
-                "--" + cls.name() + "-" + image_detail_key,
-                action="store_true",
-                dest=cls.option_name() + "_" + image_detail_key,
-                help="stack {} override".format(image_detail_key),
-            )
-
-
-#
-# Elastic Services
-#
-class ApmServer(StackService, Service):
-    docker_path = "apm"
-
-    SERVICE_PORT = "8200"
-    DEFAULT_MONITOR_PORT = "6060"
-    DEFAULT_OUTPUT = "elasticsearch"
-    OUTPUTS = {"elasticsearch", "file", "kafka", "logstash"}
-    DEFAULT_KIBANA_HOST = "kibana:5601"
-
-    def __init__(self, **options):
-        super(ApmServer, self).__init__(**options)
-
-        v1_rate_limit = ("apm-server.frontend.rate_limit", "100000")
-        if self.at_least_version("6.5"):
-            rum_config = [
-                ("apm-server.rum.enabled", "true"),
-                ("apm-server.rum.event_rate.limit", "1000"),
-            ]
-            if not self.at_least_version("7.0"):
-                rum_config.append(v1_rate_limit)
-        else:
-            rum_config = [
-                ("apm-server.frontend.enabled", "true"),
-                v1_rate_limit,
-            ]
-
-        self.apm_server_command_args = rum_config
-        self.apm_server_command_args.extend([
-            ("apm-server.host", "0.0.0.0:8200"),
-            ("apm-server.read_timeout", "1m"),
-            ("apm-server.shutdown_timeout", "2m"),
-            ("apm-server.write_timeout", "1m"),
-            ("logging.json", "true"),
-            ("logging.metrics.enabled", "false"),
-            ("setup.kibana.host", self.DEFAULT_KIBANA_HOST),
-            ("setup.template.settings.index.number_of_replicas", "0"),
-            ("setup.template.settings.index.number_of_shards", "1"),
-            ("setup.template.settings.index.refresh_interval", "1ms"),
-            ("xpack.monitoring.elasticsearch", "true"),
-            ("xpack.monitoring.enabled", "true")
-        ])
-        if options.get("apm_server_self_instrument"):
-            self.apm_server_command_args.append(("apm-server.instrumentation.enabled", "true"))
-        self.depends_on = {"elasticsearch": {"condition": "service_healthy"}} if options.get(
-            "enable_elasticsearch", True) else {}
-        self.build = self.options.get("apm_server_build")
-
-        if self.options.get("apm_server_ilm_disable"):
-            self.apm_server_command_args.append(("apm-server.ilm.enabled", "false"))
-        elif self.at_least_version("7.2") and not self.at_least_version("7.3") and not self.oss:
-            self.apm_server_command_args.append(("apm-server.ilm.enabled", "true"))
-
-        if self.options.get("apm_server_acm_disable"):
-            self.apm_server_command_args.append(("apm-server.kibana.enabled", "false"))
-        elif self.at_least_version("7.3"):
-            self.apm_server_command_args.extend([
-                ("apm-server.kibana.enabled", "true"),
-                ("apm-server.kibana.host", self.DEFAULT_KIBANA_HOST)])
-
-        if self.options.get("enable_kibana", True):
-            self.depends_on["kibana"] = {"condition": "service_healthy"}
-            if options.get("apm_server_dashboards", True) and not self.at_least_version("7.0") \
-                    and not self.options.get("xpack_secure"):
-                self.apm_server_command_args.append(
-                    ("setup.dashboards.enabled", "true")
-                )
-
-        if self.options.get("apm_server_secret_token"):
-            self.apm_server_command_args.append(("apm-server.secret_token", self.options["apm_server_secret_token"]))
-
-        self.apm_server_monitor_port = options.get("apm_server_monitor_port", self.DEFAULT_MONITOR_PORT)
-        self.apm_server_output = options.get("apm_server_output", self.DEFAULT_OUTPUT)
-
-        if options.get("apm_server_queue", "mem") == "file":
-            # enable file spool queue
-            q = {"file": {"path": "$${path.data}/spool.dat"}, "write": {}}
-            # override defaults
-            if options.get("apm_server_queue_file_size"):
-                q["file"]["size"] = options["apm_server_queue_file_size"]
-            if options.get("apm_server_queue_file_page_size"):
-                q["file"]["page_size"] = options["apm_server_queue_file_page_size"]
-            if options.get("apm_server_queue_write_buffer_size"):
-                q["write"]["buffer_size"] = options["apm_server_queue_write_buffer_size"]
-            if options.get("apm_server_queue_write_flush_events"):
-                q["write"]["flush.events"] = options["apm_server_queue_write_flush_events"]
-            if options.get("apm_server_queue_write_flush_timeout"):
-                q["write"]["flush.timeout"] = options["apm_server_queue_write_flush_timeout"]
-            if not q["write"]:
-                q.pop("write")
-            self.apm_server_command_args.append(("queue.spool", json.dumps(q)))
-
-        es_urls = []
-
-        def add_es_config(args, prefix="output"):
-            """add elasticsearch configuration options."""
-            default_apm_server_creds = {"username": "apm_server_user", "password": "changeme"}
-            args.append((prefix + ".elasticsearch.hosts", json.dumps(es_urls)))
-            for cfg in ("username", "password"):
-                es_opt = "apm_server_elasticsearch_{}".format(cfg)
-                if self.options.get(es_opt):
-                    args.append((prefix + ".elasticsearch.{}".format(cfg), self.options[es_opt]))
-                elif self.options.get("xpack_secure"):
-                    args.append((prefix + ".elasticsearch.{}".format(cfg), default_apm_server_creds.get(cfg)))
-
-        es_urls = self.options.get("apm_server_elasticsearch_urls") or [self.DEFAULT_ELASTICSEARCH_HOSTS]
-
-        if self.apm_server_output == "elasticsearch":
-            add_es_config(self.apm_server_command_args)
-            self.apm_server_command_args.extend([
-                ("output.elasticsearch.enabled", "true"),
-            ])
-            if options.get("apm_server_enable_pipeline", True) and self.at_least_version("6.5"):
-                pipeline_name = "apm" if self.at_least_version("7.2") else "apm_user_agent"
-                self.apm_server_command_args.extend([
-                    ("output.elasticsearch.pipelines", "[{pipeline: '%s'}]" % pipeline_name),
-                    ("apm-server.register.ingest.pipeline.enabled", "true"),
-                ])
-        else:
-            add_es_config(self.apm_server_command_args, prefix="xpack.monitoring")
-            self.apm_server_command_args.extend([
-                ("output.elasticsearch.enabled", "false"),
-            ])
-            if self.apm_server_output == "kafka":
-                self.apm_server_command_args.extend([
-                    ("output.kafka.enabled", "true"),
-                    ("output.kafka.hosts", "[\"kafka:9092\"]"),
-                    ("output.kafka.topics", "[{default: 'apm', topic: 'apm-%{[service.name]}'}]"),
-                ])
-            elif self.apm_server_output == "logstash":
-                self.apm_server_command_args.extend([
-                    ("output.logstash.enabled", "true"),
-                    ("output.logstash.hosts", "[\"logstash:5044\"]"),
-                ])
-            elif self.apm_server_output == "file":
-                self.apm_server_command_args.extend([
-                    ("output.file.enabled", "true"),
-                    ("output.file.path", self.options.get("apm_server_output_file", os.devnull)),
-                ])
-
-        for opt in options.get("apm_server_opt", []):
-            self.apm_server_command_args.append(opt.split("=", 1))
-
-        self.apm_server_count = options.get("apm_server_count", 1)
-        apm_server_record = options.get("apm_server_record", False)
-        self.apm_server_tee = options.get("apm_server_tee") or apm_server_record  # tee if requested or if recording
-        if self.apm_server_tee:
-            # convenience for tee without count
-            if self.apm_server_count == 1:
-                self.apm_server_count = 2
-            if apm_server_record:
-                self.apm_server_tee_build = {
-                    "context": "docker/apm-server/recorder",
-                }
-            else:
-                # always build 8.0
-                self.apm_server_tee_build = {
-                    "args": {
-                        "apm_server_base_image": "docker.elastic.co/apm/apm-server:8.0.0-SNAPSHOT",
-                        "apm_server_branch": "master",
-                        "apm_server_repo": "https://github.com/elastic/apm-server.git"
-                    },
-                    "context": "docker/apm-server"
-                }
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(ApmServer, cls).add_arguments(parser)
-        parser.add_argument(
-            '--apm-server-build',
-            const="https://github.com/elastic/apm-server.git",
-            nargs="?",
-            help='build apm-server from a git repo[@branch|sha], eg https://github.com/elastic/apm-server.git@v2'
-        )
-        parser.add_argument(
-            "--apm-server-ilm-disable",
-            action="store_true",
-            help='disable ILM (enabled by default in 7.2+)'
-        )
-        parser.add_argument(
-            '--apm-server-output',
-            choices=cls.OUTPUTS,
-            default='elasticsearch',
-            help='apm-server output'
-        )
-        parser.add_argument(
-            '--apm-server-output-file',
-            default=os.devnull,
-            help='apm-server output path (when output=file)'
-        )
-        parser.add_argument(
-            "--no-apm-server-pipeline",
-            action="store_false",
-            dest="apm_server_enable_pipeline",
-            help='disable apm-server pipelines.'
-        )
-        parser.add_argument(
-            "--no-apm-server-self-instrument",
-            action="store_false",
-            dest="apm_server_self_instrument",
-            help='disable apm-server self instrumentation.'
-        )
-        parser.add_argument(
-            '--apm-server-count',
-            type=int,
-            default=1,
-            help="apm-server count. >1 adds a load balancer service to distribute traffic between servers.",
-        )
-        parser.add_argument(
-            '--apm-server-elasticsearch-url',
-            action="append",
-            dest="apm_server_elasticsearch_urls",
-            help="apm-server elasticsearch output url(s)."
-        )
-        parser.add_argument(
-            '--apm-server-elasticsearch-username',
-            help="apm-server elasticsearch output username.",
-        )
-        parser.add_argument(
-            '--apm-server-elasticsearch-password',
-            help="apm-server elasticsearch output password.",
-        )
-        parser.add_argument(
-            "--apm-server-queue",
-            choices=("file", "mem"),
-            default="mem",
-            help="apm-server queue type.",
-        )
-        parser.add_argument(
-            "--apm-server-queue-file-size",
-            help="apm-server file spool size (eg 128MiB).",
-        )
-        parser.add_argument(
-            "--apm-server-queue-file-page-size",
-            help="apm-server file spool page size (eg 16KiB).",
-        )
-        parser.add_argument(
-            "--apm-server-queue-write-buffer-size",
-            help="apm-server file write buffer size (eg 10MiB).",
-        )
-        parser.add_argument(
-            "--apm-server-queue-write-codec",
-            choices=("cbor", "json"),
-            default="cbor",
-            help="apm-server file write codec.",
-        )
-        parser.add_argument(
-            "--apm-server-queue-write-flush-events",
-            help="apm-server file write flush event count.",
-        )
-        parser.add_argument(
-            "--apm-server-queue-write-flush-timeout",
-            help="apm-server file write flush timeout.",
-        )
-        parser.add_argument(
-            '--apm-server-secret-token',
-            dest="apm_server_secret_token",
-            help="apm-server secret token.",
-        )
-        parser.add_argument(
-            "--no-apm-server-dashboards",
-            action="store_false",
-            dest="apm_server_dashboards",
-            help="skip loading apm-server dashboards (setup.dashboards.enabled=false)",
-        )
-        parser.add_argument(
-            '--apm-server-record',
-            action="store_true",
-            default=False,
-            help=argparse.SUPPRESS,
-            # help="record apm-server request payloads.",
-        )
-        parser.add_argument(
-            '--apm-server-tee',
-            action="store_true",
-            default=False,
-            help=argparse.SUPPRESS,
-            # help="tee proxied traffic instead of load balancing.",
-        )
-        parser.add_argument(
-            "--apm-server-opt",
-            action="append",
-            default=[],
-            help="arbitrary additional configuration to set for apm-server"
-        )
-        parser.add_argument(
-            "--apm-server-acm-disable",
-            action="store_true",
-            help="disable Agent Config Management",
-        )
-
-    def build_candidate_manifest(self):
-        version = self.version
-        image = self.docker_name
-        if self.oss:
-            image += "-oss"
-
-        key = "{image}-{version}-docker-image.tar.gz".format(
-            image=image,
-            version=version,
-        )
-        try:
-            if key not in self.bc["projects"][self.docker_name]["packages"]:
-                key = "{image}-{version}-linux-amd64-docker-image.tar.gz".format(
-                    image=image,
-                    version=version,
-                )
-            return self.bc["projects"][self.docker_name]["packages"][key]
-        except KeyError:
-            # help debug manifest issues
-            print(json.dumps(self.bc))
-            raise
-
-    def _content(self):
-        command_args = []
-        for param, value in self.apm_server_command_args:
-            command_args.extend(["-E", param + "=" + value])
-
-        healthcheck_path = "/" if self.at_least_version("6.5") else "/healthcheck"
-        content = dict(
-            cap_add=["CHOWN", "DAC_OVERRIDE", "SETGID", "SETUID"],
-            cap_drop=["ALL"],
-            command=["apm-server", "-e", "--httpprof", ":{}".format(self.apm_server_monitor_port)] + command_args,
-            depends_on=self.depends_on,
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, path=healthcheck_path),
-            labels=["co.elastic.apm.stack-version=" + self.version],
-            ports=[
-                self.publish_port(self.port, self.SERVICE_PORT),
-                self.publish_port(self.apm_server_monitor_port, self.DEFAULT_MONITOR_PORT),
-            ]
-        )
-
-        if self.build:
-            build_spec_parts = self.build.split("@", 1)
-            repo = build_spec_parts[0]
-            branch = build_spec_parts[1] if len(build_spec_parts) > 1 else "master"
-            content.update({
-                "build": {
-                    "context": "docker/apm-server",
-                    "args": {
-                        "apm_server_base_image": self.default_image(),
-                        "apm_server_branch_or_commit": branch,
-                        "apm_server_repo": repo,
-                    }
-                },
-                "image": None,
-            })
-
-        return content
-
-    @staticmethod
-    def enabled():
-        return True
-
-    def render(self):
-        """hack up render to support multiple apm servers behind a load balancer"""
-        ren = super(ApmServer, self).render()
-        if self.apm_server_count == 1:
-            return ren
-
-        # save a single server for use as backend template
-        single = ren[self.name()]
-        single["ports"] = [p.rsplit(":", 1)[-1] for p in single["ports"]]
-
-        # render proxy + backends
-        if self.apm_server_tee:
-            ren = self.render_tee()
-        else:
-            ren = self.render_proxy()
-
-        # individualize each backend instance
-        for i in range(1, self.apm_server_count + 1):
-            backend = dict(single)
-            backend["container_name"] = backend["container_name"] + "-" + str(i)
-            if self.apm_server_tee and i > 1:
-                backend["build"] = self.apm_server_tee_build
-                backend["labels"] = ["co.elastic.apm.stack-version=8.0.0"]
-                del(backend["image"])  # use the built one instead
-            ren.update({"-".join([self.name(), str(i)]): backend})
-
-        return ren
-
-    def render_proxy(self):
-        condition = {"condition": "service_healthy"}
-        content = dict(
-            build={"context": "docker/apm-server/haproxy"},
-            container_name=self.default_container_name() + "-load-balancer",
-            depends_on={"apm-server-{}".format(i): condition for i in range(1, self.apm_server_count + 1)},
-            environment={"APM_SERVER_COUNT": self.apm_server_count},
-            healthcheck={"test": ["CMD", "haproxy", "-c", "-f", "/usr/local/etc/haproxy/haproxy.cfg"]},
-            ports=[
-                self.publish_port(self.port, self.SERVICE_PORT),
-            ],
-        )
-        return {self.name(): content}
-
-    def render_tee(self):
-        condition = {"condition": "service_healthy"}
-        command = ["teeproxy", "-l", "0.0.0.0:8200", "-a", "apm-server-1:8200", "-b", "apm-server-2:8200"]
-        # add extra tee backends
-        for i in range(3, self.apm_server_count + 1):
-            command.extend(["-b", self.name() + "-" + str(i) + ":8200"])
-        content = dict(
-            build={"context": "docker/apm-server/teeproxy"},
-            command=command,
-            container_name=self.default_container_name() + "-tee",
-            depends_on={"apm-server-{}".format(i): condition for i in range(1, self.apm_server_count + 1)},
-            healthcheck={"test": ["CMD", "pgrep", "teeproxy"]},
-            ports=[
-                self.publish_port(self.port, self.SERVICE_PORT),
-            ],
-        )
-        return {self.name(): content}
-
-
-class Elasticsearch(StackService, Service):
-    default_environment = [
-        "bootstrap.memory_lock=true",
-        "cluster.name=docker-cluster",
-        "cluster.routing.allocation.disk.threshold_enabled=false",
-        "discovery.type=single-node",
-        "path.repo=/usr/share/elasticsearch/data/backups",
-    ]
-    default_heap_size = "1g"
-
-    SERVICE_PORT = 9200
-
-    def __init__(self, **options):
-        super(Elasticsearch, self).__init__(**options)
-        if not self.oss and not self.at_least_version("6.3"):
-            self.docker_name = self.name() + "-platinum"
-
-        self.xpack_secure = bool(self.options.get("xpack_secure"))
-
-        # construct elasticsearch environment variables
-        es_java_opts = {}
-        if "elasticsearch_heap" in options:
-            es_java_opts.update({
-                "Xms": options["elasticsearch_heap"],
-                "Xmx": options["elasticsearch_heap"],
-            })
-        es_java_opts.update(dict(options.get("elasticsearch_java_opts", {}) or {}))
-        if self.at_least_version("6.4"):
-            # per https://github.com/elastic/elasticsearch/pull/32138/files
-            es_java_opts["XX:UseAVX"] = "=2"
-
-        java_opts_env = "ES_JAVA_OPTS=" + " ".join(["-{}{}".format(k, v) for k, v in sorted(es_java_opts.items())])
-        # falsy empty string permitted
-        data_dir = self.version if options.get("elasticsearch_data_dir") is None else options["elasticsearch_data_dir"]
-
-        self.environment = self.default_environment + [
-            java_opts_env, "path.data=/usr/share/elasticsearch/data/" + data_dir]
-        if not self.oss:
-            xpack_security_enabled = "false"
-            if self.xpack_secure:
-                xpack_security_enabled = "true"
-                if options.get("elasticsearch_xpack_audit"):
-                    self.environment.append("xpack.security.audit.enabled=true")
-                self.environment.append("xpack.security.authc.anonymous.roles=remote_monitoring_collector")
-                if self.at_least_version("7.0"):
-                    self.environment.append("xpack.security.authc.realms.file.file1.order=0")
-                    self.environment.append("xpack.security.authc.realms.native.native1.order=1")
-                else:
-                    self.environment.append("xpack.security.authc.realms.file1.type=file")
-                    self.environment.append("xpack.security.authc.realms.native1.type=native")
-                    self.environment.append("xpack.security.authc.realms.native1.order=1")
-            self.environment.append("xpack.security.enabled=" + xpack_security_enabled)
-            self.environment.append("xpack.license.self_generated.type=trial")
-            if self.at_least_version("6.3"):
-                self.environment.append("xpack.monitoring.collection.enabled=true")
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(Elasticsearch, cls).add_arguments(parser)
-        parser.add_argument(
-            "--elasticsearch-data-dir",
-            help="override elasticsearch data dir.  Defaults to the current es version."
-        )
-
-        parser.add_argument(
-            "--elasticsearch-heap",
-            default=Elasticsearch.default_heap_size,
-            help="min/max elasticsearch heap size, for -Xms -Xmx jvm options."
-        )
-
-        parser.add_argument(
-            "--elasticsearch-xpack-audit",
-            action="store_true",
-            help="enable very verbose xpack auditing",
-        )
-
-        class storeDict(argparse.Action):
-            def __call__(self, parser, namespace, value, option_string=None):
-                items = getattr(namespace, self.dest)
-                items.update(dict([value.split("=", 1)]))
-
-        # this is a dict to enable deduplication
-        # eg --elasticsearch-java-opts a==z --elasticsearch-java-opts a==b will add only -a=b to ES_JAVA_OPTS
-        parser.add_argument(
-            "--elasticsearch-java-opts",
-            action=storeDict,
-            default={},
-            help="additional entries for ES_JAVA_OPTS, multiple allowed, separate key/value with =."
-        )
-
-    def _content(self):
-        volumes = ["esdata:/usr/share/elasticsearch/data"]
-        if self.xpack_secure:
-            volumes.extend([
-                "./docker/elasticsearch/roles.yml:/usr/share/elasticsearch/config/roles.yml",
-                "./docker/elasticsearch/users:/usr/share/elasticsearch/config/users",
-                "./docker/elasticsearch/users_roles:/usr/share/elasticsearch/config/users_roles",
-            ])
-        return dict(
-            environment=self.environment,
-            healthcheck={
-                "interval": "20",
-                "retries": 10,
-                "test": ["CMD-SHELL", "curl -s http://localhost:9200/_cluster/health | grep -vq '\"status\":\"red\"'"],
-            },
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-            ulimits={
-                "memlock": {"hard": -1, "soft": -1},
-            },
-            volumes=volumes,
-        )
-
-    @staticmethod
-    def enabled():
-        return True
-
-
-class BeatMixin(object):
-    DEFAULT_OUTPUT = "elasticsearch"
-    OUTPUTS = {"elasticsearch", "logstash"}
-
-    @classmethod
-    def add_arguments(cls, parser):
-        parser.add_argument(
-            "--{}-elasticsearch-url".format(cls.name()),
-            action="append",
-            dest="{}_elasticsearch_urls".format(cls.name()),
-            help="{} elasticsearch output url(s).".format(cls.name())
-        )
-        parser.add_argument(
-            "--{}-elasticsearch-username".format(cls.name()),
-            help="{} elasticsearch output username.".format(cls.name()),
-        )
-        parser.add_argument(
-            "--{}-elasticsearch-password".format(cls.name()),
-            help="{} elasticsearch output password.".format(cls.name()),
-        )
-        parser.add_argument(
-            "--{}-output".format(cls.name()),
-            choices=cls.OUTPUTS,
-            default="elasticsearch",
-            help="{} output".format(cls.name()),
-        )
-
-    def __init__(self, **options):
-        super(BeatMixin, self).__init__(**options)
-        self.command = list(self.DEFAULT_COMMAND)
-        self.depends_on = {"elasticsearch": {"condition": "service_healthy"}} if options.get(
-            "enable_elasticsearch", True) else {}
-        if options.get("enable_kibana", True):
-            self.command.extend(["-E", "setup.dashboards.enabled=true"])
-            self.depends_on["kibana"] = {"condition": "service_healthy"}
-        self.environment = {}
-
-        def add_es_config(args, prefix="output"):
-            """add elasticsearch configuration options."""
-            es_urls = options.get("{}_elasticsearch_urls".format(self.name())) or [self.DEFAULT_ELASTICSEARCH_HOSTS]
-            default_beat_creds = {"username": "{}_user".format(self.name()), "password": "changeme"}
-            args.append((prefix + ".elasticsearch.hosts", json.dumps(es_urls)))
-            for cfg in ("username", "password"):
-                es_opt = "{}_elasticsearch_{}".format(self.name(), cfg)
-                if options.get(es_opt):
-                    args.append((prefix + ".elasticsearch.{}".format(cfg), options[es_opt]))
-                elif options.get("xpack_secure"):
-                    args.append((prefix + ".elasticsearch.{}".format(cfg), default_beat_creds.get(cfg)))
-
-        command_args = []
-        add_es_config(command_args)
-        beat_output = options.get("{}_output".format(self.name()), self.DEFAULT_OUTPUT)
-        if beat_output == "elasticsearch":
-            command_args.extend([("output.elasticsearch.enabled", "true")])
-        else:
-            command_args.extend([("output.elasticsearch.enabled", "false")])
-            add_es_config(command_args, prefix="xpack.monitoring")
-            if beat_output == "logstash":
-                command_args.extend([
-                    ("output.logstash.enabled", "true"),
-                    ("output.logstash.hosts", "[\"logstash:5044\"]"),
-                ])
-            elif beat_output == "kafka":
-                # disabled via command line options for now
-                command_args.extend([
-                    ("output.kafka.enabled", "true"),
-                    ("output.kafka.hosts", "[\"kafka:9092\"]"),
-                    ("output.kafka.topics", "[{default: '{}', topic: '{}'}]".format(self.name(), self.name())),
-                ])
-
-        for param, value in command_args:
-            self.command.extend(["-E", param + "=" + value])
-
-    def build_candidate_manifest(self):
-        version = self.version
-        image = self.docker_name
-        if self.oss:
-            image += "-oss"
-
-        key = "{image}-{version}-docker-image.tar.gz".format(
-            image=image,
-            version=version,
-        )
-        try:
-            if key not in self.bc["projects"]["beats"]["packages"]:
-                # This is the old standard based on the format:
-                # ${name}-${version}-${os}-${architecture}-${classifier}.${extension}
-                key = "{image}-{version}-linux-amd64-docker-image.tar.gz".format(
-                    image=image,
-                    version=version,
-                )
-            return self.bc["projects"]["beats"]["packages"][key]
-        except KeyError:
-            try:
-                # This is the new standard based on the format:
-                # ${name}-${version}-${classifier}-${os}-${architecture}.${extension}
-                key = "{image}-{version}-docker-image-linux-amd64.tar.gz".format(
-                        image=image,
-                        version=version,
-                )
-                return self.bc["projects"]["beats"]["packages"][key]
-            except KeyError:
-                # help debug manifest issues
-                print(json.dumps(self.bc))
-                raise
-
-
-class Filebeat(BeatMixin, StackService, Service):
-    DEFAULT_COMMAND = ["filebeat", "-e", "--strict.perms=false"]
-    docker_path = "beats"
-
-    def __init__(self, **options):
-        super(Filebeat, self).__init__(**options)
-        config = "filebeat.yml" if self.at_least_version("6.1") else "filebeat.simple.yml"
-        self.filebeat_config_path = os.path.join(".", "docker", "filebeat", config)
-
-    def _content(self):
-        return dict(
-            command=self.command,
-            depends_on=self.depends_on,
-            environment=self.environment,
-            labels=None,
-            user="root",
-            volumes=[
-                self.filebeat_config_path + ":/usr/share/filebeat/filebeat.yml",
-                "/var/lib/docker/containers:/var/lib/docker/containers",
-                "/var/run/docker.sock:/var/run/docker.sock",
-            ]
-        )
-
-
-class Heartbeat(BeatMixin, StackService, Service):
-    DEFAULT_COMMAND = ["heartbeat", "-e", "--strict.perms=false"]
-    docker_path = "beats"
-
-    def __init__(self, **options):
-        options['enable_kibana'] = False
-        super(Heartbeat, self).__init__(**options)
-        config = "heartbeat.yml"
-        self.heartbeat_config_path = os.path.join(".", "docker", "heartbeat", config)
-
-    def _content(self):
-        return dict(
-            command=self.command,
-            depends_on=self.depends_on,
-            environment=self.environment,
-            labels=None,
-            user="root",
-            volumes=[
-                self.heartbeat_config_path + ":/usr/share/heartbeat/heartbeat.yml",
-                "/var/lib/docker/containers:/var/lib/docker/containers",
-                "/var/run/docker.sock:/var/run/docker.sock",
-            ]
-        )
-
-
-class Kibana(StackService, Service):
-    default_environment = {"SERVER_NAME": "kibana.example.org"}
-
-    SERVICE_PORT = 5601
-
-    def __init__(self, **options):
-        super(Kibana, self).__init__(**options)
-        if not self.at_least_version("6.3") and not self.oss:
-            self.docker_name = self.name() + "-x-pack"
-        self.environment = self.default_environment.copy()
-        if not self.oss:
-            self.environment["XPACK_MONITORING_ENABLED"] = "true"
-            if self.at_least_version("6.3"):
-                self.environment["XPACK_XPACK_MAIN_TELEMETRY_ENABLED"] = "false"
-            if options.get("xpack_secure"):
-                self.environment["ELASTICSEARCH_PASSWORD"] = "changeme"
-                self.environment["ELASTICSEARCH_USERNAME"] = "kibana_system_user"
-                self.environment["STATUS_ALLOWANONYMOUS"] = "true"
-        self.environment["ELASTICSEARCH_URL"] = ",".join(self.options.get(
-            "kibana_elasticsearch_urls") or [self.DEFAULT_ELASTICSEARCH_HOSTS])
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(Kibana, cls).add_arguments(parser)
-        parser.add_argument(
-            "--{}-elasticsearch-url".format(cls.name()),
-            action="append",
-            dest="kibana_elasticsearch_urls",
-            help="kibana elasticsearch output url(s)."
-        )
-
-    def _content(self):
-        return dict(
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "kibana", path="/api/status", retries=20),
-            depends_on={"elasticsearch": {"condition": "service_healthy"}} if self.options.get(
-                "enable_elasticsearch", True) else {},
-            environment=self.environment,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-    @staticmethod
-    def enabled():
-        return True
-
-
-class Logstash(StackService, Service):
-    SERVICE_PORT = 5044
-
-    def build_candidate_manifest(self):
-        version = self.version
-        image = self.docker_name
-        if self.oss:
-            image += "-oss"
-        key = "{image}-{version}-docker-image.tar.gz".format(
-            image=image,
-            version=version,
-        )
-        return self.bc["projects"]["logstash-docker"]["packages"][key]
-
-    def _content(self):
-        self.es_urls = ",".join(self.options.get(
-            "logstash_elasticsearch_urls") or [self.DEFAULT_ELASTICSEARCH_HOSTS])
-        return dict(
-            depends_on={"elasticsearch": {"condition": "service_healthy"}} if self.options.get(
-                "enable_elasticsearch", True) else {},
-            environment={
-                "ELASTICSEARCH_URL": self.es_urls,
-                },
-            healthcheck=curl_healthcheck(9600, "logstash", path="/"),
-            ports=[self.publish_port(self.port, self.SERVICE_PORT), "9600"],
-            volumes=["./docker/logstash/pipeline/:/usr/share/logstash/pipeline/"]
-        )
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(Logstash, cls).add_arguments(parser)
-        parser.add_argument(
-            "--{}-elasticsearch-url".format(cls.name()),
-            action="append",
-            dest="logstash_elasticsearch_urls",
-            help="logstash elasticsearch output url(s)."
-        )
-
-
-class Metricbeat(BeatMixin, StackService, Service):
-    DEFAULT_COMMAND = ["metricbeat", "-e", "--strict.perms=false"]
-    docker_path = "beats"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(Metricbeat, cls).add_arguments(parser)
-        parser.add_argument(
-            "--apm-server-pprof-url",
-            help="apm server profiling url to use.",
-            dest="apm_server_pprof_url",
-            default="apm-server:6060"
-        )
-
-    def _content(self):
-        self.environment['APM_SERVER_PPROF_HOST'] = self.options.get("apm_server_pprof_url")
-        return dict(
-            command=self.command,
-            depends_on=self.depends_on,
-            environment=self.environment,
-            labels=None,
-            user="root",
-            volumes=[
-                "./docker/metricbeat/metricbeat.yml:/usr/share/metricbeat/metricbeat.yml",
-                "/var/run/docker.sock:/var/run/docker.sock",
-            ]
-        )
-
-
-class Packetbeat(BeatMixin, StackService, Service):
-    """Stars a Packetbeat container to grab the network traffic."""
-
-    DEFAULT_COMMAND = ["packetbeat", "-e", "--strict.perms=false", "-E", "packetbeat.interfaces.device=eth0"]
-    docker_path = "beats"
-
-    def _content(self):
-        return dict(
-            command=self.command,
-            depends_on=self.depends_on,
-            environment=self.environment,
-            labels=None,
-            user="root",
-            privileged="true",
-            cap_add=["NET_ADMIN", "NET_RAW"],
-            network_mode="service:apm-server",
-            volumes=[
-                "./docker/packetbeat/packetbeat.yml:/usr/share/packetbeat/packetbeat.yml",
-                "/var/run/docker.sock:/var/run/docker.sock",
-            ]
-        )
-
-
-#
-# Supporting Services
-#
-class Kafka(Service):
-    SERVICE_PORT = 9092
-
-    def _content(self):
-        return dict(
-            depends_on=["zookeeper"],
-            environment={
-                "KAFKA_ADVERTISED_LISTENERS": "PLAINTEXT://kafka:9092",
-                "KAFKA_BROKER_ID": 1,
-                "KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR": 1,
-                "KAFKA_ZOOKEEPER_CONNECT": "zookeeper:2181",
-            },
-            image="confluentinc/cp-kafka:4.1.3",
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class Postgres(Service):
-    SERVICE_PORT = 5432
-    opbeans_side_car = True
-
-    def _content(self):
-        return dict(
-            environment=["POSTGRES_DB=opbeans", "POSTGRES_PASSWORD=verysecure"],
-            healthcheck={"interval": "10s", "test": ["CMD", "pg_isready", "-h", "postgres", "-U", "postgres"]},
-            image="postgres:10",
-            labels=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT, expose=True)],
-            volumes=["./docker/opbeans/sql:/docker-entrypoint-initdb.d", "pgdata:/var/lib/postgresql/data"],
-        )
-
-
-class Redis(Service):
-    SERVICE_PORT = 6379
-    opbeans_side_car = True
-
-    def _content(self):
-        return dict(
-            healthcheck={"interval": "10s", "test": ["CMD", "redis-cli", "ping"]},
-            image="redis:4",
-            labels=None,
-            command="--save ''",  # disable persistence
-            ports=[self.publish_port(self.port, self.SERVICE_PORT, expose=True)],
-        )
-
-
-class Zookeeper(Service):
-    SERVICE_PORT = 2181
-
-    def _content(self):
-        return dict(
-            environment={
-                "ZOOKEEPER_CLIENT_PORT": 2181,
-                "ZOOKEEPER_TICK_TIME": 2000,
-            },
-            image="confluentinc/cp-zookeeper:latest",
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-#
-# Agent Integration Test Services
-#
-class AgentRUMJS(Service):
-    SERVICE_PORT = 8000
-    DEFAULT_AGENT_BRANCH = "master"
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-rum-js"
-
-    def __init__(self, **options):
-        super(AgentRUMJS, self).__init__(**options)
-        self.agent_branch = options.get("rum_agent_branch", self.DEFAULT_AGENT_BRANCH)
-        self.agent_repo = options.get("rum_agent_repo", self.DEFAULT_AGENT_REPO)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentRUMJS, cls).add_arguments(parser)
-        parser.add_argument(
-            '--rum-agent-repo',
-            default=cls.DEFAULT_AGENT_REPO,
-            help="GitHub repo to be used. Default: {}".format(cls.DEFAULT_AGENT_REPO),
-        )
-        parser.add_argument(
-            '--rum-agent-branch',
-            default=cls.DEFAULT_AGENT_BRANCH,
-        )
-
-    def _content(self):
-        return dict(
-            build=dict(
-                context="docker/rum",
-                dockerfile="Dockerfile",
-                args=[
-                    "RUM_AGENT_BRANCH=" + self.agent_branch,
-                    "RUM_AGENT_REPO=" + self.agent_repo,
-                    "APM_SERVER_URL=" + self.options.get("apm_server_url", DEFAULT_APM_SERVER_URL)
-                ]
-            ),
-            container_name="rum",
-            image=None,
-            labels=None,
-            logging=None,
-            environment={
-                "ELASTIC_APM_SERVICE_NAME": "rum",
-                "ELASTIC_APM_SERVER_URL": self.options.get("apm_server_url", DEFAULT_APM_SERVER_URL)
-            },
-            depends_on=self.depends_on,
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "rum", path="/"),
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentGoNetHttp(Service):
-    SERVICE_PORT = 8080
-    DEFAULT_AGENT_VERSION = "master"
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-go"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentGoNetHttp, cls).add_arguments(parser)
-        parser.add_argument(
-            "--go-agent-version",
-            default=cls.DEFAULT_AGENT_VERSION,
-            help='Use Go agent version (master, 0.5, v0.5.2, ...)',
-        )
-        parser.add_argument(
-            '--go-agent-repo',
-            default=cls.DEFAULT_AGENT_REPO,
-            help="GitHub repo to be used. Default: {}".format(cls.DEFAULT_AGENT_REPO),
-        )
-
-    def __init__(self, **options):
-        super(AgentGoNetHttp, self).__init__(**options)
-        self.agent_version = options.get("go_agent_version", self.DEFAULT_AGENT_VERSION)
-        self.agent_repo = options.get("go_agent_repo", self.DEFAULT_AGENT_REPO)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "ELASTIC_APM_SERVER_URL"),
-    ])
-    def _content(self):
-        return dict(
-            build={
-                "context": "docker/go/nethttp",
-                "dockerfile": "Dockerfile",
-                "args": {
-                    "GO_AGENT_BRANCH": self.agent_version,
-                    "GO_AGENT_REPO": self.agent_repo,
-                },
-            },
-            container_name="gonethttpapp",
-            environment={
-                "ELASTIC_APM_API_REQUEST_TIME": "3s",
-                "ELASTIC_APM_FLUSH_INTERVAL": "500ms",
-                "ELASTIC_APM_SERVICE_NAME": "gonethttpapp",
-                "ELASTIC_APM_TRANSACTION_IGNORE_NAMES": "healthcheck",
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "gonethttpapp"),
-            depends_on=self.depends_on,
-            image=None,
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentNodejsExpress(Service):
-    # elastic/apm-agent-nodejs#master
-    DEFAULT_AGENT_PACKAGE = "elastic-apm-node"
-    SERVICE_PORT = 8010
-
-    def __init__(self, **options):
-        super(AgentNodejsExpress, self).__init__(**options)
-        self.agent_package = options.get("nodejs_agent_package", self.DEFAULT_AGENT_PACKAGE)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentNodejsExpress, cls).add_arguments(parser)
-        parser.add_argument(
-            '--nodejs-agent-package',
-            default=cls.DEFAULT_AGENT_PACKAGE,
-            help='Use Node.js agent version (github;master, github;1.x, github;v3.0.0, release;latest, ...)',
-        )
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "ELASTIC_APM_SERVER_URL"),
-    ])
-    def _content(self):
-        return dict(
-            build={"context": "docker/nodejs/express", "dockerfile": "Dockerfile"},
-            command="bash -c \"npm install {} && node app.js\"".format(
-                self.agent_package, self.SERVICE_PORT),
-            container_name="expressapp",
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "expressapp"),
-            depends_on=self.depends_on,
-            image=None,
-            labels=None,
-            logging=None,
-            environment={
-                "EXPRESS_PORT": str(self.SERVICE_PORT),
-                "EXPRESS_SERVICE_NAME": "expressapp",
-            },
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentPython(Service):
-    DEFAULT_AGENT_PACKAGE = "elastic-apm"
-    _agent_python_arguments_added = False
-
-    def __init__(self, **options):
-        super(AgentPython, self).__init__(**options)
-        self.agent_package = options.get("python_agent_package", self.DEFAULT_AGENT_PACKAGE)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @classmethod
-    def add_arguments(cls, parser):
-        if cls._agent_python_arguments_added:
-            return
-
-        super(AgentPython, cls).add_arguments(parser)
-        parser.add_argument(
-            '--python-agent-package',
-            default=cls.DEFAULT_AGENT_PACKAGE,
-            help='Use Python agent version (github;master, github;1.x, github;v3.0.0, release;latest, ...)',
-        )
-        # prevent calling again
-        cls._agent_python_arguments_added = True
-
-    def _content(self):
-        raise NotImplementedError()
-
-
-class AgentPythonDjango(AgentPython):
-    SERVICE_PORT = 8003
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "APM_SERVER_URL"),
-    ])
-    def _content(self):
-        return dict(
-            build={"context": "docker/python/django", "dockerfile": "Dockerfile"},
-            command="bash -c \"pip install -q -U {} && python testapp/manage.py runserver 0.0.0.0:{}\"".format(
-                self.agent_package, self.SERVICE_PORT),
-            container_name="djangoapp",
-            environment={
-                "DJANGO_PORT": self.SERVICE_PORT,
-                "DJANGO_SERVICE_NAME": "djangoapp",
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "djangoapp"),
-            depends_on=self.depends_on,
-            image=None,
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentPythonFlask(AgentPython):
-    SERVICE_PORT = 8001
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "APM_SERVER_URL"),
-    ])
-    def _content(self):
-        return dict(
-            build={"context": "docker/python/flask", "dockerfile": "Dockerfile"},
-            command="bash -c \"pip install -q -U {} && gunicorn app:app\"".format(self.agent_package),
-            container_name="flaskapp",
-            image=None,
-            labels=None,
-            logging=None,
-            environment={
-                "FLASK_SERVICE_NAME": "flaskapp",
-                "GUNICORN_CMD_ARGS": "-w 4 -b 0.0.0.0:{}".format(self.SERVICE_PORT),
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "flaskapp"),
-            depends_on=self.depends_on,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentRubyRails(Service):
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-ruby"
-    DEFAULT_AGENT_VERSION = "latest"
-    DEFAULT_AGENT_VERSION_STATE = "release"
-    SERVICE_PORT = 8020
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentRubyRails, cls).add_arguments(parser)
-        parser.add_argument(
-            "--ruby-agent-version",
-            default=cls.DEFAULT_AGENT_VERSION,
-            help='Use Ruby agent version (master, 1.x, latest, ...)',
-        )
-        parser.add_argument(
-            "--ruby-agent-version-state",
-            default=cls.DEFAULT_AGENT_VERSION_STATE,
-            help='Use Ruby agent version state (github or release)',
-        )
-        parser.add_argument(
-            "--ruby-agent-repo",
-            default=cls.DEFAULT_AGENT_REPO,
-            help="GitHub repo to be used. Default: {}".format(cls.DEFAULT_AGENT_REPO),
-        )
-
-    def __init__(self, **options):
-        super(AgentRubyRails, self).__init__(**options)
-        self.agent_version = options.get("ruby_agent_version", self.DEFAULT_AGENT_VERSION)
-        self.agent_version_state = options.get("ruby_agent_version_state", self.DEFAULT_AGENT_VERSION_STATE)
-        self.agent_repo = options.get("ruby_agent_repo", self.DEFAULT_AGENT_REPO)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-    ])
-    def _content(self):
-        return dict(
-            build={
-                "context": "docker/ruby/rails",
-                "dockerfile": "Dockerfile",
-                "args": {
-                    "RUBY_AGENT_VERSION": self.agent_version,
-                    "RUBY_AGENT_REPO": self.agent_repo,
-                }
-            },
-            command="bash -c \"bundle install && RAILS_ENV=production bundle exec rails s -b 0.0.0.0 -p {}\"".format(
-                self.SERVICE_PORT),
-            container_name="railsapp",
-            environment={
-                "APM_SERVER_URL": self.options.get("apm_server_url", DEFAULT_APM_SERVER_URL),
-                "ELASTIC_APM_API_REQUEST_TIME": "3s",
-                "ELASTIC_APM_SERVER_URL": self.options.get("apm_server_url", DEFAULT_APM_SERVER_URL),
-                "ELASTIC_APM_SERVICE_NAME": "railsapp",
-                "RAILS_PORT": self.SERVICE_PORT,
-                "RAILS_SERVICE_NAME": "railsapp",
-                "RUBY_AGENT_VERSION_STATE": self.agent_version_state,
-                "RUBY_AGENT_VERSION": self.agent_version,
-                "RUBY_AGENT_REPO": self.agent_repo,
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "railsapp", retries=60),
-            depends_on=self.depends_on,
-            image=None,
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentJavaSpring(Service):
-    SERVICE_PORT = 8090
-    DEFAULT_AGENT_VERSION = "master"
-    DEFAULT_AGENT_RELEASE = ""
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-java"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentJavaSpring, cls).add_arguments(parser)
-        parser.add_argument(
-            "--java-agent-version",
-            default=cls.DEFAULT_AGENT_VERSION,
-            help='Use Java agent version (master, 0.5, v.0.7.1, ...)',
-        )
-        parser.add_argument(
-            "--java-agent-release",
-            default=cls.DEFAULT_AGENT_RELEASE,
-            help='Use Java agent release version (1.6.0, 0.6.2, ...)',
-        )
-        parser.add_argument(
-            "--java-agent-repo",
-            default=cls.DEFAULT_AGENT_REPO,
-            help="GitHub repo to be used. Default: {}".format(cls.DEFAULT_AGENT_REPO),
-        )
-
-    def __init__(self, **options):
-        super(AgentJavaSpring, self).__init__(**options)
-        self.agent_version = options.get("java_agent_version", self.DEFAULT_AGENT_VERSION)
-        self.agent_release = options.get("java_agent_release", self.DEFAULT_AGENT_RELEASE)
-        self.agent_repo = options.get("java_agent_repo", self.DEFAULT_AGENT_REPO)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "ELASTIC_APM_SERVER_URL"),
-    ])
-    def _content(self):
-        return dict(
-            build={
-                "context": "docker/java/spring",
-                "dockerfile": "Dockerfile",
-                "args": {
-                    "JAVA_AGENT_BRANCH": self.agent_version,
-                    "JAVA_AGENT_BUILT_VERSION": self.agent_release,
-                    "JAVA_AGENT_REPO": self.agent_repo,
-                }
-            },
-            container_name="javaspring",
-            image=None,
-            labels=None,
-            logging=None,
-            environment={
-                "ELASTIC_APM_API_REQUEST_TIME": "3s",
-                "ELASTIC_APM_SERVICE_NAME": "springapp",
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "javaspring"),
-            depends_on=self.depends_on,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-class AgentDotnet(Service):
-    SERVICE_PORT = 8100
-    DEFAULT_AGENT_VERSION = "master"
-    DEFAULT_AGENT_RELEASE = ""
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-dotnet"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(AgentDotnet, cls).add_arguments(parser)
-        parser.add_argument(
-            "--dotnet-agent-version",
-            default=cls.DEFAULT_AGENT_VERSION,
-            help='Use .NET agent version (master, 0.0.0.2, 0.0.0.1, ...)',
-        )
-        parser.add_argument(
-            "--dotnet-agent-release",
-            default=cls.DEFAULT_AGENT_RELEASE,
-            help='Use .NET agent release version (0.0.1-alpha, 0.0.2-alpha, ...)',
-        )
-        parser.add_argument(
-            "--dotnet-agent-repo",
-            default=cls.DEFAULT_AGENT_REPO,
-            help="GitHub repo to be used. Default: {}".format(cls.DEFAULT_AGENT_REPO),
-        )
-
-    def __init__(self, **options):
-        super(AgentDotnet, self).__init__(**options)
-        self.agent_version = options.get("dotnet_agent_version", self.DEFAULT_AGENT_VERSION)
-        self.agent_release = options.get("dotnet_agent_release", self.DEFAULT_AGENT_RELEASE)
-        self.agent_repo = options.get("dotnet_agent_repo", self.DEFAULT_AGENT_REPO)
-        if options.get("enable_apm_server", True):
-            self.depends_on = {
-                "apm-server": {"condition": "service_healthy"},
-            }
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN"),
-        ("apm_server_url", "ELASTIC_APM_SERVER_URLS"),
-    ])
-    def _content(self):
-        return dict(
-            build={
-                "context": "docker/dotnet",
-                "dockerfile": "Dockerfile",
-                "args": {
-                    "DOTNET_AGENT_BRANCH": self.agent_version,
-                    "DOTNET_AGENT_VERSION": self.agent_release,
-                    "DOTNET_AGENT_REPO": self.agent_repo,
-                },
-            },
-            container_name="dotnetapp",
-            environment={
-                "ELASTIC_APM_API_REQUEST_TIME": "3s",
-                "ELASTIC_APM_FLUSH_INTERVAL": "5",
-                "ELASTIC_APM_SAMPLE_RATE": "1",
-                "ELASTIC_APM_SERVICE_NAME": "dotnetapp",
-                "ELASTIC_APM_TRANSACTION_IGNORE_NAMES": "healthcheck",
-            },
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, "dotnetapp"),
-            depends_on=self.depends_on,
-            image=None,
-            labels=None,
-            logging=None,
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-
-
-#
-# Opbeans Services
-#
-
-
-class OpbeansService(Service):
-    def __init__(self, **options):
-        super(OpbeansService, self).__init__(**options)
-        self.apm_server_url = options.get("apm_server_url", DEFAULT_APM_SERVER_URL)
-        self.apm_js_server_url = options.get("opbeans_apm_js_server_url", DEFAULT_APM_JS_SERVER_URL)
-        self.opbeans_dt_probability = options.get("opbeans_dt_probability", 0.5)
-        if hasattr(self, "DEFAULT_SERVICE_NAME"):
-            self.service_name = options.get(self.option_name() + "_service_name", self.DEFAULT_SERVICE_NAME)
-        self.agent_branch = options.get(self.option_name() + "_agent_branch") or ""
-        self.agent_repo = options.get(self.option_name() + "_agent_repo") or ""
-        self.agent_local_repo = options.get(self.option_name() + "_agent_local_repo")
-        self.opbeans_branch = options.get(self.option_name() + "_branch") or ""
-        self.opbeans_repo = options.get(self.option_name() + "_repo") or ""
-        self.es_urls = ",".join(self.options.get("opbeans_elasticsearch_urls") or [self.DEFAULT_ELASTICSEARCH_HOSTS])
-        self.service_environment = \
-            options.get(self.option_name() + "_service_environment") or self.DEFAULT_ELASTIC_APM_ENVIRONMENT
-
-    @classmethod
-    def add_arguments(cls, parser):
-        """add service-specific command line arguments"""
-        # allow port overrides
-        super(OpbeansService, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-agent-branch',
-            default=None,
-            dest=cls.option_name() + '_agent_branch',
-            help=cls.name() + " branch for agent"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-agent-repo',
-            default=None,
-            dest=cls.option_name() + '_agent_repo',
-            help=cls.name() + " github repo for agent (in form org/repo)"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-agent-local-repo',
-            default=None,
-            dest=cls.option_name() + '_agent_local_repo',
-            help=cls.name() + " local repo path for agent"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-service-environment',
-            default=None,
-            dest=cls.option_name() + '_service_environment',
-            help=cls.name() + " service.environment value to display."
-        )
-        if hasattr(cls, 'DEFAULT_SERVICE_NAME'):
-            parser.add_argument(
-                '--' + cls.name() + '-service-name',
-                default=cls.DEFAULT_SERVICE_NAME,
-                dest=cls.option_name() + '_service_name',
-                help=cls.name() + " service name"
-            )
-
-
-class OpbeansDotnet(OpbeansService):
-    SERVICE_PORT = 3004
-    DEFAULT_AGENT_BRANCH = "master"
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-dotnet"
-    DEFAULT_SERVICE_NAME = "opbeans-dotnet"
-    DEFAULT_AGENT_VERSION = ""
-    DEFAULT_OPBEANS_BRANCH = "master"
-    DEFAULT_OPBEANS_REPO = "elastic/opbeans-dotnet"
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansDotnet, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-version',
-            default=cls.DEFAULT_AGENT_VERSION,
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-branch',
-            default=cls.DEFAULT_OPBEANS_BRANCH,
-            dest=cls.option_name() + '_branch',
-            help=cls.name() + " branch for the opbeans dotnet"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-repo',
-            default=cls.DEFAULT_OPBEANS_REPO,
-            dest=cls.option_name() + '_repo',
-            help=cls.name() + " github repo for the opbeans dotnet (in form org/repo)"
-        )
-
-    def __init__(self, **options):
-        super(OpbeansDotnet, self).__init__(**options)
-        self.agent_version = options.get('opbeans_dotnet_version')
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {}
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-        if self.options.get("enable_elasticsearch", True):
-            depends_on["elasticsearch"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/dotnet",
-                dockerfile="Dockerfile",
-                args=[
-                    "DOTNET_AGENT_BRANCH=" + (self.agent_branch or self.DEFAULT_AGENT_BRANCH),
-                    "DOTNET_AGENT_REPO=" + (self.agent_repo or self.DEFAULT_AGENT_REPO),
-                    "DOTNET_AGENT_VERSION=" + (self.agent_version or self.DEFAULT_AGENT_VERSION),
-                    "OPBEANS_DOTNET_BRANCH=" + (self.opbeans_branch or self.DEFAULT_OPBEANS_BRANCH),
-                    "OPBEANS_DOTNET_REPO=" + (self.opbeans_repo or self.DEFAULT_OPBEANS_REPO),
-                ]
-            ),
-            environment=[
-                "ELASTIC_APM_SERVICE_NAME={}".format(self.service_name),
-                "ELASTIC_APM_SERVER_URLS={}".format(self.apm_server_url),
-                "ELASTIC_APM_JS_SERVER_URL={}".format(self.apm_js_server_url),
-                "ELASTIC_APM_FLUSH_INTERVAL=5",
-                "ELASTIC_APM_TRANSACTION_MAX_SPANS=50",
-                "ELASTIC_APM_SAMPLE_RATE=1",
-                "ELASTICSEARCH_URL={}".format(self.es_urls),
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            healthcheck=curl_healthcheck(80, "opbeans-dotnet", path="/", retries=36),
-            ports=[self.publish_port(self.port, 80)],
-        )
-        return content
-
-
-class OpbeansDotnet01(OpbeansDotnet):
-    SERVICE_PORT = 3104
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansDotnet01, self).__init__(**options)
-
-
-class OpbeansGo(OpbeansService):
-    SERVICE_PORT = 3003
-    DEFAULT_AGENT_BRANCH = "master"
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-go"
-    DEFAULT_OPBEANS_BRANCH = "master"
-    DEFAULT_OPBEANS_REPO = "elastic/opbeans-go"
-    DEFAULT_SERVICE_NAME = "opbeans-go"
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansGo, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-branch',
-            default=cls.DEFAULT_OPBEANS_BRANCH,
-            dest=cls.option_name() + '_branch',
-            help=cls.name() + " branch for the opbeans go"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-repo',
-            default=cls.DEFAULT_OPBEANS_REPO,
-            dest=cls.option_name() + '_repo',
-            help=cls.name() + " github repo for the opbeans go (in form org/repo)"
-        )
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {
-            "postgres": {"condition": "service_healthy"},
-            "redis": {"condition": "service_healthy"},
-        }
-
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-        if self.options.get("enable_elasticsearch", True):
-            depends_on["elasticsearch"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/go",
-                dockerfile="Dockerfile",
-                args=[
-                    "GO_AGENT_BRANCH=" + (self.agent_branch or self.DEFAULT_AGENT_BRANCH),
-                    "GO_AGENT_REPO=" + (self.agent_repo or self.DEFAULT_AGENT_REPO),
-                    "OPBEANS_GO_BRANCH=" + (self.opbeans_branch or self.DEFAULT_OPBEANS_BRANCH),
-                    "OPBEANS_GO_REPO=" + (self.opbeans_repo or self.DEFAULT_OPBEANS_REPO),
-                ]
-            ),
-            environment=[
-                "ELASTIC_APM_SERVICE_NAME={}".format(self.service_name),
-                "ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "ELASTIC_APM_JS_SERVER_URL={}".format(self.apm_js_server_url),
-                "ELASTIC_APM_FLUSH_INTERVAL=5",
-                "ELASTIC_APM_TRANSACTION_MAX_SPANS=50",
-                "ELASTIC_APM_SAMPLE_RATE=1",
-                "ELASTICSEARCH_URL={}".format(self.es_urls),
-                "OPBEANS_CACHE=redis://redis:6379",
-                "OPBEANS_PORT=3000",
-                "PGHOST=postgres",
-                "PGPORT=5432",
-                "PGUSER=postgres",
-                "PGPASSWORD=verysecure",
-                "PGSSLMODE=disable",
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            ports=[self.publish_port(self.port, 3000)],
-        )
-        return content
-
-
-class OpbeansGo01(OpbeansGo):
-    SERVICE_PORT = 3103
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansGo01, self).__init__(**options)
-
-
-class OpbeansJava(OpbeansService):
-    SERVICE_PORT = 3002
-    DEFAULT_AGENT_BRANCH = ""
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-java"
-    DEFAULT_LOCAL_REPO = "."
-    DEFAULT_SERVICE_NAME = 'opbeans-java'
-    DEFAULT_OPBEANS_IMAGE = 'opbeans/opbeans-java'
-    DEFAULT_OPBEANS_VERSION = 'latest'
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansJava, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-image',
-            default=cls.DEFAULT_OPBEANS_IMAGE,
-            help=cls.name() + " image for the opbeans java"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-version',
-            default=cls.DEFAULT_OPBEANS_VERSION,
-            help=cls.name() + " version for the docker image of opbeans java"
-        )
-
-    def __init__(self, **options):
-        super(OpbeansJava, self).__init__(**options)
-        self.opbeans_image = options.get('opbeans_java_image')
-        self.opbeans_version = options.get('opbeans_java_version')
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {
-            "postgres": {"condition": "service_healthy"},
-        }
-
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-        if self.options.get("enable_elasticsearch", True):
-            depends_on["elasticsearch"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/java",
-                dockerfile="Dockerfile",
-                args=[
-                    "JAVA_AGENT_BRANCH=" + (self.agent_branch or self.DEFAULT_AGENT_BRANCH),
-                    "JAVA_AGENT_REPO=" + (self.agent_repo or self.DEFAULT_AGENT_REPO),
-                    "OPBEANS_JAVA_IMAGE=" + (self.opbeans_image or self.DEFAULT_OPBEANS_IMAGE),
-                    "OPBEANS_JAVA_VERSION=" + (self.opbeans_version or self.DEFAULT_OPBEANS_VERSION),
-                ]
-            ),
-            environment=[
-                "ELASTIC_APM_SERVICE_NAME={}".format(self.service_name),
-                "ELASTIC_APM_APPLICATION_PACKAGES=co.elastic.apm.opbeans",
-                "ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "ELASTIC_APM_FLUSH_INTERVAL=5",
-                "ELASTIC_APM_TRANSACTION_MAX_SPANS=50",
-                "ELASTIC_APM_SAMPLE_RATE=1",
-                "ELASTIC_APM_ENABLE_LOG_CORRELATION=true",
-                "DATABASE_URL=jdbc:postgresql://postgres/opbeans?user=postgres&password=verysecure",
-                "DATABASE_DIALECT=POSTGRESQL",
-                "DATABASE_DRIVER=org.postgresql.Driver",
-                "REDIS_URL=redis://redis:6379",
-                "ELASTICSEARCH_URL={}".format(self.es_urls),
-                "OPBEANS_SERVER_PORT=3000",
-                "JAVA_AGENT_VERSION",
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            healthcheck=curl_healthcheck(3000, "opbeans-java", path="/", retries=36),
-            ports=[self.publish_port(self.port, 3000)],
-        )
-        if self.agent_local_repo:
-            content["volumes"] = [
-                "{}:/local-install".format(self.agent_local_repo),
-            ]
-        return content
-
-
-class OpbeansJava01(OpbeansJava):
-    SERVICE_PORT = 3102
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansJava01, self).__init__(**options)
-
-
-class OpbeansNode(OpbeansService):
-    SERVICE_PORT = 3000
-    DEFAULT_LOCAL_REPO = "."
-    DEFAULT_OPBEANS_IMAGE = 'opbeans/opbeans-node'
-    DEFAULT_OPBEANS_VERSION = 'latest'
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansNode, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-image',
-            default=cls.DEFAULT_OPBEANS_IMAGE,
-            help=cls.name() + " image for the opbeans node"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-version',
-            default=cls.DEFAULT_OPBEANS_VERSION,
-            help=cls.name() + " version for the docker image of opbeans node"
-        )
-
-    def __init__(self, **options):
-        super(OpbeansNode, self).__init__(**options)
-        self.service_name = "opbeans-node"
-        self.opbeans_image = options.get('opbeans_node_image')
-        self.opbeans_version = options.get('opbeans_node_version')
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {
-            "postgres": {"condition": "service_healthy"},
-            "redis": {"condition": "service_healthy"},
-        }
-
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/node",
-                dockerfile="Dockerfile",
-                args=[
-                    "OPBEANS_NODE_IMAGE=" + (self.opbeans_image or self.DEFAULT_OPBEANS_IMAGE),
-                    "OPBEANS_NODE_VERSION=" + (self.opbeans_version or self.DEFAULT_OPBEANS_VERSION),
-                ]
-            ),
-            environment=[
-                "ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "ELASTIC_APM_JS_SERVER_URL={}".format(self.apm_js_server_url),
-                "ELASTIC_APM_LOG_LEVEL=info",
-                "ELASTIC_APM_SOURCE_LINES_ERROR_APP_FRAMES",
-                "ELASTIC_APM_SOURCE_LINES_SPAN_APP_FRAMES=5",
-                "ELASTIC_APM_SOURCE_LINES_ERROR_LIBRARY_FRAMES",
-                "ELASTIC_APM_SOURCE_LINES_SPAN_LIBRARY_FRAMES",
-                "WORKLOAD_ELASTIC_APM_APP_NAME=workload",
-                "WORKLOAD_ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "WORKLOAD_DISABLED={}".format(self.options.get("no_opbeans_node_loadgen", False)),
-                "OPBEANS_SERVER_PORT=3000",
-                "OPBEANS_SERVER_HOSTNAME=opbeans-node",
-                "NODE_ENV=production",
-                "PGHOST=postgres",
-                "PGPASSWORD=verysecure",
-                "PGPORT=5432",
-                "PGUSER=postgres",
-                "REDIS_URL=redis://redis:6379",
-                "NODE_AGENT_BRANCH=" + self.agent_branch,
-                "NODE_AGENT_REPO=" + self.agent_repo,
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            healthcheck=curl_healthcheck(3000, "opbeans-node", path="/"),
-            ports=[self.publish_port(self.port, 3000)],
-            volumes=[
-                "./docker/opbeans/node/sourcemaps:/sourcemaps",
-            ]
-        )
-        if self.agent_local_repo:
-            content["volumes"].append(
-                "{}:/local-install".format(self.agent_local_repo),
-            )
-        return content
-
-
-class OpbeansNode01(OpbeansNode):
-    SERVICE_PORT = 3100
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansNode01, self).__init__(**options)
-
-
-class OpbeansPython(OpbeansService):
-    SERVICE_PORT = 8000
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-python"
-    DEFAULT_AGENT_BRANCH = "2.x"
-    DEFAULT_LOCAL_REPO = "."
-    DEFAULT_SERVICE_NAME = 'opbeans-python'
-    DEFAULT_OPBEANS_IMAGE = 'opbeans/opbeans-python'
-    DEFAULT_OPBEANS_VERSION = 'latest'
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansPython, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-local-repo',
-            default=cls.DEFAULT_LOCAL_REPO,
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-image',
-            default=cls.DEFAULT_OPBEANS_IMAGE,
-            help=cls.name() + " image for the opbeans python"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-version',
-            default=cls.DEFAULT_OPBEANS_VERSION,
-            help=cls.name() + " version for the docker image of opbeans python"
-        )
-
-    def __init__(self, **options):
-        super(OpbeansPython, self).__init__(**options)
-        self.opbeans_image = options.get('opbeans_python_image')
-        self.opbeans_version = options.get('opbeans_python_version')
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {
-            "postgres": {"condition": "service_healthy"},
-            "redis": {"condition": "service_healthy"},
-        }
-
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-        if self.options.get("enable_elasticsearch", True):
-            depends_on["elasticsearch"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/python",
-                dockerfile="Dockerfile",
-                args=[
-                    "OPBEANS_PYTHON_IMAGE=" + (self.opbeans_image or self.DEFAULT_OPBEANS_IMAGE),
-                    "OPBEANS_PYTHON_VERSION=" + (self.opbeans_version or self.DEFAULT_OPBEANS_VERSION),
-                ]
-            ),
-            environment=[
-                "DATABASE_URL=postgres://postgres:verysecure@postgres/opbeans",
-                "ELASTIC_APM_SERVICE_NAME={}".format(self.service_name),
-                "ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "ELASTIC_APM_JS_SERVER_URL={}".format(self.apm_js_server_url),
-                "ELASTIC_APM_FLUSH_INTERVAL=5",
-                "ELASTIC_APM_TRANSACTION_MAX_SPANS=50",
-                "ELASTIC_APM_TRANSACTION_SAMPLE_RATE=0.5",
-                "ELASTIC_APM_SOURCE_LINES_ERROR_APP_FRAMES",
-                "ELASTIC_APM_SOURCE_LINES_SPAN_APP_FRAMES=5",
-                "ELASTIC_APM_SOURCE_LINES_ERROR_LIBRARY_FRAMES",
-                "ELASTIC_APM_SOURCE_LINES_SPAN_LIBRARY_FRAMES",
-                "REDIS_URL=redis://redis:6379",
-                "ELASTICSEARCH_URL={}".format(self.es_urls),
-                "OPBEANS_SERVER_URL=http://opbeans-python:3000",
-                "PYTHON_AGENT_BRANCH=" + self.agent_branch,
-                "PYTHON_AGENT_REPO=" + self.agent_repo,
-                "PYTHON_AGENT_VERSION",
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            healthcheck=curl_healthcheck(3000, "opbeans-python", path="/"),
-            ports=[self.publish_port(self.port, 3000)],
-        )
-        if self.agent_local_repo:
-            content["volumes"] = [
-                "{}:/local-install".format(self.agent_local_repo),
-            ]
-        return content
-
-
-class OpbeansPython01(OpbeansPython):
-    SERVICE_PORT = 8100
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansPython01, self).__init__(**options)
-
-
-class OpbeansRuby(OpbeansService):
-    SERVICE_PORT = 3001
-    DEFAULT_AGENT_BRANCH = "master"
-    DEFAULT_AGENT_REPO = "elastic/apm-agent-ruby"
-    DEFAULT_LOCAL_REPO = "."
-    DEFAULT_SERVICE_NAME = "opbeans-ruby"
-    DEFAULT_OPBEANS_IMAGE = 'opbeans/opbeans-ruby'
-    DEFAULT_OPBEANS_VERSION = 'latest'
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "production"
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansRuby, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-image',
-            default=cls.DEFAULT_OPBEANS_IMAGE,
-            help=cls.name() + " image for the opbeans ruby"
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-version',
-            default=cls.DEFAULT_OPBEANS_VERSION,
-            help=cls.name() + " version for the docker image of opbeans ruby"
-        )
-
-    def __init__(self, **options):
-        super(OpbeansRuby, self).__init__(**options)
-        self.opbeans_image = options.get('opbeans_ruby_image')
-        self.opbeans_version = options.get('opbeans_ruby_version')
-
-    @add_agent_environment([
-        ("apm_server_secret_token", "ELASTIC_APM_SECRET_TOKEN")
-    ])
-    def _content(self):
-        depends_on = {
-            "postgres": {"condition": "service_healthy"},
-            "redis": {"condition": "service_healthy"},
-        }
-
-        if self.options.get("enable_apm_server", True):
-            depends_on["apm-server"] = {"condition": "service_healthy"}
-        if self.options.get("enable_elasticsearch", True):
-            depends_on["elasticsearch"] = {"condition": "service_healthy"}
-
-        content = dict(
-            build=dict(
-                context="docker/opbeans/ruby",
-                dockerfile="Dockerfile",
-                args=[
-                    "OPBEANS_RUBY_IMAGE=" + (self.opbeans_image or self.DEFAULT_OPBEANS_IMAGE),
-                    "OPBEANS_RUBY_VERSION=" + (self.opbeans_version or self.DEFAULT_OPBEANS_VERSION),
-                ]
-            ),
-            environment=[
-                "ELASTIC_APM_SERVER_URL={}".format(self.apm_server_url),
-                "ELASTIC_APM_SERVICE_NAME={}".format(self.service_name),
-                "DATABASE_URL=postgres://postgres:verysecure@postgres/opbeans-ruby",
-                "REDIS_URL=redis://redis:6379",
-                "ELASTICSEARCH_URL={}".format(self.es_urls),
-                "OPBEANS_SERVER_URL=http://opbeans-ruby:3000",
-                "RAILS_ENV=production",
-                "RAILS_LOG_TO_STDOUT=1",
-                "PORT=3000",
-                "RUBY_AGENT_BRANCH=" + self.agent_branch,
-                "RUBY_AGENT_REPO=" + self.agent_repo,
-                "RUBY_AGENT_VERSION",
-                "OPBEANS_DT_PROBABILITY={:.2f}".format(self.opbeans_dt_probability),
-                "ELASTIC_APM_ENVIRONMENT={}".format(self.service_environment),
-            ],
-            depends_on=depends_on,
-            image=None,
-            labels=None,
-            # lots of retries as the ruby app can take a long time to boot
-            healthcheck=curl_healthcheck(3000, "opbeans-ruby", path="/", retries=50),
-            ports=[self.publish_port(self.port, 3000)],
-        )
-        if self.agent_local_repo:
-            content["volumes"] = [
-                "{}:/local-install".format(self.agent_local_repo),
-            ]
-        return content
-
-
-class OpbeansRuby01(OpbeansRuby):
-    SERVICE_PORT = 3101
-    DEFAULT_ELASTIC_APM_ENVIRONMENT = "testing"
-
-    def __init__(self, **options):
-        super(OpbeansRuby01, self).__init__(**options)
-
-
-class OpbeansRum(Service):
-    # OpbeansRum is not really an Opbeans service, so we inherit from Service
-    SERVICE_PORT = 9222
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansRum, cls).add_arguments(parser)
-        parser.add_argument(
-            '--' + cls.name() + '-backend-service',
-            default='opbeans-node',
-        )
-        parser.add_argument(
-            '--' + cls.name() + '-backend-port',
-            default='3000',
-        )
-
-    def __init__(self, **options):
-        super(OpbeansRum, self).__init__(**options)
-        self.backend_service = options.get('opbeans_rum_backend_service', 'opbeans-node')
-        self.backend_port = options.get('opbeans_rum_backend_port', '3000')
-
-    def _content(self):
-        content = dict(
-            build={"context": "docker/opbeans/rum", "dockerfile": "Dockerfile"},
-            cap_add=["SYS_ADMIN"],
-            depends_on={self.backend_service: {'condition': 'service_healthy'}},
-            environment=[
-                "OPBEANS_BASE_URL=http://{}:{}".format(self.backend_service, self.backend_port),
-            ],
-            image=None,
-            labels=None,
-            healthcheck=curl_healthcheck(self.SERVICE_PORT, path="/"),
-            ports=[self.publish_port(self.port, self.SERVICE_PORT)],
-        )
-        return content
-
-
-class OpbeansLoadGenerator(Service):
-    opbeans_side_car = True
-
-    @classmethod
-    def add_arguments(cls, parser):
-        super(OpbeansLoadGenerator, cls).add_arguments(parser)
-        for service_class in OpbeansService.__subclasses__():
-            parser.add_argument(
-                '--no-%s-loadgen' % service_class.name(),
-                action='store_true',
-                default=False,
-                help='Disable load generator for {}'.format(service_class.name())
-            )
-            parser.add_argument(
-                '--%s-loadgen-rpm' % service_class.name(),
-                action='store',
-                default=100,
-                help='RPM of load that should be generated for {}'.format(service_class.name())
-            )
-
-    def __init__(self, **options):
-        super(OpbeansLoadGenerator, self).__init__(**options)
-        self.loadgen_services = []
-        self.loadgen_rpms = OrderedDict()
-        # create load for opbeans services
-        run_all_opbeans = options.get('run_all_opbeans') or options.get('run_all')
-        excluded = ('opbeans_load_generator', 'opbeans_rum', 'opbeans_node', 'opbeans_node', 'opbeans_dotnet01',
-                    'opbeans_go01', 'opbeans_java01', 'opbeans_node01', 'opbeans_python01', 'opbeans_ruby01')
-        for flag, value in options.items():
-            if (value or run_all_opbeans) and flag.startswith('enable_opbeans_'):
-                service_name = flag[len('enable_'):]
-                if not options.get('no_{}_loadgen'.format(service_name)) and service_name not in excluded:
-                    self.loadgen_services.append(service_name.replace('_', '-'))
-                    rpm = options.get('{}_loadgen_rpm'.format(service_name))
-                    if rpm:
-                        self.loadgen_rpms[service_name.replace('_', '-')] = rpm
-
-    def _content(self):
-        content = dict(
-            image="opbeans/opbeans-loadgen:latest",
-            depends_on={service: {'condition': 'service_healthy'} for service in self.loadgen_services},
-            environment=[
-                "OPBEANS_URLS={}".format(','.join('{0}:http://{0}:3000'.format(s) for s in sorted(self.loadgen_services))),  # noqa: E501
-                "OPBEANS_RPMS={}".format(','.join('{}:{}'.format(k, v) for k, v in sorted(self.loadgen_rpms.items())))
-            ],
-            labels=None,
-        )
-        return content
-
 
 #
 # Service Tests
